third_party/blink/web_tests/platform/generic/editing/pasteboard/paste-blockquote-before-blockquote-expected.txt

CONSOLE MESSAGE: SUCCESS: pasting singleNestedNodeToCopy DID NOT keep the newline within the blockquote.
This tests pasting a singly nested blockquote before, and outside of, another blockquote. This test does not work outside of DRT.

line 2


line 1
line 2
line 3
line 4
line 5
